backend/pyproject.toml
[project]
name = "lvlup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
